refactor(board): refactor chunked upload handlers for better error handling
- Extract helper functions for room loading, reservation validation, and
  chunk processing in complete.rs
- Add proper error handling and validation for all steps of file merging
- Introduce structured validation functions for reservations and chunks
- Use separate functions for file operations like merging, verification,
  and cleanup
- Replace inline logic with descriptive function calls for better
  readability
- Add proper error propagation using dedicated validation functions

feat(board): enhance chunk upload handler with structured parsing

- Create ParsedChunkUpload struct to organize upload data
- Implement separate functions for field parsing and validation
- Add proper error handling for each step of chunk processing
- Extract validation functions for reservation checking and room
  matching
- Implement separate functions for file writing and database
  persistence
