+++ crates/board/src/handlers/chunked_upload/complete.rs
@@ -1,3 +1,4 @@
+use std::path::Path as StdPath;
 use std::sync::Arc;
 
 use axum::{
@@ -15,9 +16,10 @@ use crate::{
     dto::chunked_upload::{FileMergeRequest, FileMergeResponse, MergedFileInfo},
     errors::{AppError, AppResult},
     models::room::{
+        Room,
         chunk_upload::RoomChunkUpload,
         content::{ContentType, RoomContent},
-        upload_reservation::{UploadFileDescriptor, UploadStatus},
+        upload_reservation::{RoomUploadReservation, UploadFileDescriptor, UploadStatus},
     },
     repository::{
         IRoomContentRepository, IRoomRepository,
@@ -60,55 +62,154 @@ pub async fn complete_file_merge(
     RoomNameValidator::validate_identifier(&room_name)?;
 
     let room_repository = RoomRepository::new(app_state.db_pool.clone());
-    let room = room_repository
-        .find_by_name(&room_name)
-        .await
-        .map_err(|e| AppError::internal(format!("数据库查询错误：{}", e)))?;
-    let room = room.ok_or_else(|| AppError::not_found("房间不存在"))?;
-
+    let room = load_room(&room_repository, &room_name).await?;
+    let room_id = room
+        .id
+        .ok_or_else(|| AppError::internal("房间 ID 不能为空"))?;
     let reservation_repository = RoomUploadReservationRepository::new(app_state.db_pool.clone());
-    let reservation_id: i64 = payload
-        .reservation_id
-        .parse()
-        .map_err(|_| AppError::validation("reservation_id 必须是数字"))?;
-    let reservation = reservation_repository
+    let reservation_id = parse_reservation_id(&payload.reservation_id)?;
+    let reservation = load_reservation(&reservation_repository, reservation_id).await?;
+    validate_reservation_for_merge(&reservation, room_id)?;
+
+    let chunk_repository = RoomChunkUploadRepository::new(app_state.db_pool.clone());
+    let reservation_db_id = reservation_id_or_error(&reservation)?;
+    let sorted_chunks = load_uploaded_chunks(
+        &chunk_repository,
+        reservation_db_id,
+        reservation.total_chunks,
+    )
+    .await?;
+    let temp_dir = format!("/tmp/elizabeth/chunks/{}/", payload.reservation_id);
+    let final_file_path = format!("{}/merged_file", temp_dir);
+
+    merge_and_verify_chunks(
+        &sorted_chunks,
+        &final_file_path,
+        &payload.final_hash,
+        &reservation_repository,
+        reservation_db_id,
+    )
+    .await?;
+
+    let file_manifest = parse_file_manifest(&reservation.file_manifest)?;
+    let file = first_manifest_file(&file_manifest)?;
+    let storage_dir = format!("storage/rooms/{room_id}/");
+    fs::create_dir_all(&storage_dir)
+        .await
+        .map_err(|e| AppError::internal(format!("创建存储目录失败：{}", e)))?;
+
+    let final_storage_path = unique_storage_path(&storage_dir, &file.name)?;
+    fs::rename(&final_file_path, &final_storage_path)
+        .await
+        .map_err(|e| AppError::internal(format!("移动文件失败：{}", e)))?;
+
+    reservation_repository
+        .update_upload_status(reservation_db_id, UploadStatus::Completed)
+        .await
+        .map_err(|e| AppError::internal(format!("更新上传状态失败：{}", e)))?;
+    reservation_repository
+        .consume_upload(reservation_db_id)
+        .await
+        .map_err(|e| AppError::internal(format!("消费预留记录失败：{}", e)))?;
+
+    cleanup_temp_dir(&temp_dir).await;
+
+    let content_repository = RoomContentRepository::new(app_state.db_pool.clone());
+    let created_content =
+        create_content_record(&content_repository, room_id, file, &final_storage_path).await?;
+    update_room_size(
+        &room_repository,
+        &room,
+        reservation.reserved_size,
+        file.size,
+    )
+    .await?;
+
+    Ok(Json(FileMergeResponse {
+        reservation_id: payload.reservation_id.clone(),
+        merged_files: vec![MergedFileInfo {
+            file_name: file.name.clone(),
+            file_size: file.size,
+            file_hash: payload.final_hash.clone(),
+            content_id: created_content.id,
+        }],
+        message: "文件合并完成".to_string(),
+    }))
+}
+
+async fn load_room(repository: &RoomRepository, room_name: &str) -> Result<Room, AppError> {
+    repository
+        .find_by_name(room_name)
+        .await
+        .map_err(|e| AppError::internal(format!("数据库查询错误：{}", e)))?
+        .ok_or_else(|| AppError::not_found("房间不存在"))
+}
+
+fn parse_reservation_id(raw: &str) -> Result<i64, AppError> {
+    raw.parse()
+        .map_err(|_| AppError::validation("reservation_id 必须是数字"))
+}
+
+async fn load_reservation(
+    repository: &RoomUploadReservationRepository,
+    reservation_id: i64,
+) -> Result<RoomUploadReservation, AppError> {
+    repository
         .find_by_reservation_id(reservation_id)
         .await
-        .map_err(|e| AppError::internal(format!("查询预留记录失败：{}", e)))?;
-    let reservation = reservation.ok_or_else(|| AppError::not_found("预留记录不存在"))?;
+        .map_err(|e| AppError::internal(format!("查询预留记录失败：{}", e)))?
+        .ok_or_else(|| AppError::not_found("预留记录不存在"))
+}
 
-    if reservation.room_id != room.id.expect("房间 ID 不能为空") {
+fn reservation_id_or_error(reservation: &RoomUploadReservation) -> Result<i64, AppError> {
+    reservation
+        .id
+        .ok_or_else(|| AppError::internal("预留 ID 不能为空"))
+}
+
+fn validate_reservation_for_merge(
+    reservation: &RoomUploadReservation,
+    room_id: i64,
+) -> Result<(), AppError> {
+    if reservation.room_id != room_id {
         return Err(AppError::permission_denied("预留记录不属于指定房间"));
     }
-
     if reservation.chunked_upload != Some(true) {
         return Err(AppError::validation("非分块上传预留记录"));
     }
-
     if reservation.expires_at < Utc::now().naive_utc() {
         return Err(AppError::permission_denied("预留记录已过期"));
     }
-
     if reservation.upload_status == Some(UploadStatus::Completed) {
         return Err(AppError::conflict("文件已完成合并"));
     }
-
     if reservation.upload_status == Some(UploadStatus::Failed) {
         return Err(AppError::conflict("上传已失败，无法合并"));
     }
+    Ok(())
+}
 
-    let chunk_repository = RoomChunkUploadRepository::new(app_state.db_pool.clone());
-    let reservation_db_id = reservation.id.expect("预留 ID 不能为空");
-    let chunks = chunk_repository
-        .find_by_reservation_id(reservation_db_id)
+async fn load_uploaded_chunks(
+    repository: &RoomChunkUploadRepository,
+    reservation_id: i64,
+    total_chunks: Option<i64>,
+) -> Result<Vec<RoomChunkUpload>, AppError> {
+    let chunks = repository
+        .find_by_reservation_id(reservation_id)
         .await
         .map_err(|e| AppError::internal(format!("查询分块记录失败：{}", e)))?;
+    let total_chunks = total_chunks.unwrap_or(0);
+    validate_uploaded_chunks(&chunks, total_chunks)?;
 
-    let total_chunks = reservation.total_chunks.unwrap_or(0);
+    let mut sorted_chunks = chunks;
+    sorted_chunks.sort_by_key(|chunk| chunk.chunk_index);
+    Ok(sorted_chunks)
+}
+
+fn validate_uploaded_chunks(chunks: &[RoomChunkUpload], total_chunks: i64) -> Result<(), AppError> {
     if total_chunks == 0 {
         return Err(AppError::validation("总分块数为 0"));
     }
-
     if chunks.len() != total_chunks as usize {
         return Err(AppError::validation(format!(
             "分块未全部上传，已上传：{}，总分块：{}",
@@ -116,8 +217,7 @@ pub async fn complete_file_merge(
             total_chunks
         )));
     }
-
-    for chunk in &chunks {
+    for chunk in chunks {
         if !chunk.is_uploaded() {
             return Err(AppError::validation(format!(
                 "分块{}未上传完成",
@@ -125,154 +225,148 @@ pub async fn complete_file_merge(
             )));
         }
     }
+    Ok(())
+}
 
-    let mut sorted_chunks = chunks;
-    sorted_chunks.sort_by_key(|chunk| chunk.chunk_index);
+async fn merge_and_verify_chunks(
+    chunks: &[RoomChunkUpload],
+    final_file_path: &str,
+    final_hash: &str,
+    repository: &RoomUploadReservationRepository,
+    reservation_id: i64,
+) -> Result<(), AppError> {
+    if let Err(e) = merge_chunks(chunks, final_file_path).await {
+        mark_upload_failed(repository, reservation_id).await;
+        return Err(AppError::internal(format!("文件合并失败：{}", e)));
+    }
 
-    let temp_dir = format!("/tmp/elizabeth/chunks/{}/", payload.reservation_id);
-    let final_file_path = format!("{}/merged_file", temp_dir);
-
-    match merge_chunks(&sorted_chunks, &final_file_path).await {
-        Ok(_) => match verify_file_hash(&final_file_path, &payload.final_hash).await {
-            Ok(true) => {
-                let storage_dir = format!("storage/rooms/{}/", room.id.expect("房间 ID 不能为空"));
-                fs::create_dir_all(&storage_dir)
-                    .await
-                    .map_err(|e| AppError::internal(format!("创建存储目录失败：{}", e)))?;
-
-                let file_manifest: Vec<UploadFileDescriptor> =
-                    serde_json::from_str(&reservation.file_manifest)
-                        .map_err(|e| AppError::internal(format!("解析文件清单失败：{}", e)))?;
-
-                if file_manifest.is_empty() {
-                    return Err(AppError::internal("文件清单为空"));
-                }
-
-                let file_name = &file_manifest[0].name;
-                let safe_file_name = sanitize_filename::sanitize(file_name);
-                let mut final_filename = safe_file_name.clone();
-                let mut counter = 1;
-                let mut final_storage_path = format!("{}{}", storage_dir, final_filename);
-
-                while std::path::Path::new(&final_storage_path).exists() {
-                    let path = std::path::Path::new(&safe_file_name);
-                    let stem = path
-                        .file_stem()
-                        .and_then(|s| s.to_str())
-                        .unwrap_or(&safe_file_name);
-                    let extension = path.extension().and_then(|s| s.to_str()).unwrap_or("");
-
-                    final_filename = if extension.is_empty() {
-                        format!("{}({})", stem, counter)
-                    } else {
-                        format!("{}({}).{}", stem, counter, extension)
-                    };
-
-                    final_storage_path = format!("{}{}", storage_dir, final_filename);
-                    counter += 1;
-
-                    if counter > 1000 {
-                        return Err(AppError::internal("Too many files with the same name"));
-                    }
-                }
-
-                fs::rename(&final_file_path, &final_storage_path)
-                    .await
-                    .map_err(|e| AppError::internal(format!("移动文件失败：{}", e)))?;
-
-                reservation_repository
-                    .update_upload_status(reservation_db_id, UploadStatus::Completed)
-                    .await
-                    .map_err(|e| AppError::internal(format!("更新上传状态失败：{}", e)))?;
-
-                reservation_repository
-                    .consume_upload(reservation_db_id)
-                    .await
-                    .map_err(|e| AppError::internal(format!("消费预留记录失败：{}", e)))?;
-
-                if let Err(e) = fs::remove_dir_all(&temp_dir).await {
-                    logrs::error!("清理临时文件失败：{}", e);
-                }
-
-                let file_name = &file_manifest[0].name;
-                let mime_type = file_manifest[0]
-                    .mime
-                    .clone()
-                    .unwrap_or_else(|| "application/octet-stream".to_string());
-
-                let content_repository = RoomContentRepository::new(app_state.db_pool.clone());
-                let now = chrono::Utc::now().naive_utc();
-                let room_id_value: i64 = room.id.expect("房间 ID 不能为空");
-                let room_content = RoomContent {
-                    id: None,
-                    room_id: room_id_value,
-                    content_type: ContentType::File,
-                    text: None,
-                    url: Some(file_name.clone()),
-                    path: Some(final_storage_path.clone()),
-                    file_name: Some(file_name.clone()),
-                    size: Some(file_manifest[0].size),
-                    mime_type: Some(mime_type),
-                    sequence_number: 0,
-                    created_at: now,
-                    updated_at: now,
-                };
-
-                let created_content = content_repository
-                    .create(&room_content)
-                    .await
-                    .map_err(|e| AppError::internal(format!("创建内容记录失败：{}", e)))?;
-
-                let file_size = file_manifest[0].size;
-                let reserved_size = reservation.reserved_size;
-                let mut updated_room = room.clone();
-                updated_room.current_size =
-                    (updated_room.current_size - reserved_size + file_size).max(0);
-                let room_repository = RoomRepository::new(app_state.db_pool.clone());
-                room_repository
-                    .update(&updated_room)
-                    .await
-                    .map_err(|e| AppError::internal(format!("更新房间大小失败：{}", e)))?;
-
-                let merged_file_info = MergedFileInfo {
-                    file_name: file_name.clone(),
-                    file_size: file_manifest[0].size,
-                    file_hash: payload.final_hash.clone(),
-                    content_id: created_content.id,
-                };
-
-                let response = FileMergeResponse {
-                    reservation_id: payload.reservation_id.clone(),
-                    merged_files: vec![merged_file_info],
-                    message: "文件合并完成".to_string(),
-                };
-
-                Ok(Json(response))
-            }
-            Ok(false) => {
-                let _ = fs::remove_file(&final_file_path).await;
-                let _ = reservation_repository
-                    .update_upload_status(reservation_db_id, UploadStatus::Failed)
-                    .await;
-
-                Err(AppError::validation("文件哈希验证失败"))
-            }
-            Err(e) => {
-                let _ = fs::remove_file(&final_file_path).await;
-                let _ = reservation_repository
-                    .update_upload_status(reservation_db_id, UploadStatus::Failed)
-                    .await;
-
-                Err(AppError::internal(format!("文件哈希验证失败：{}", e)))
-            }
-        },
-        Err(e) => {
-            let _ = reservation_repository
-                .update_upload_status(reservation_db_id, UploadStatus::Failed)
-                .await;
-
-            Err(AppError::internal(format!("文件合并失败：{}", e)))
+    match verify_file_hash(final_file_path, final_hash).await {
+        Ok(true) => Ok(()),
+        Ok(false) => {
+            cleanup_failed_merge(final_file_path, repository, reservation_id).await;
+            Err(AppError::validation("文件哈希验证失败"))
         }
+        Err(e) => {
+            cleanup_failed_merge(final_file_path, repository, reservation_id).await;
+            Err(AppError::internal(format!("文件哈希验证失败：{}", e)))
+        }
+    }
+}
+
+async fn mark_upload_failed(repository: &RoomUploadReservationRepository, reservation_id: i64) {
+    let _ = repository
+        .update_upload_status(reservation_id, UploadStatus::Failed)
+        .await;
+}
+
+async fn cleanup_failed_merge(
+    final_file_path: &str,
+    repository: &RoomUploadReservationRepository,
+    reservation_id: i64,
+) {
+    let _ = fs::remove_file(final_file_path).await;
+    mark_upload_failed(repository, reservation_id).await;
+}
+
+fn parse_file_manifest(manifest: &str) -> Result<Vec<UploadFileDescriptor>, AppError> {
+    serde_json::from_str(manifest)
+        .map_err(|e| AppError::internal(format!("解析文件清单失败：{}", e)))
+}
+
+fn first_manifest_file(
+    file_manifest: &[UploadFileDescriptor],
+) -> Result<&UploadFileDescriptor, AppError> {
+    file_manifest
+        .first()
+        .ok_or_else(|| AppError::internal("文件清单为空"))
+}
+
+fn unique_storage_path(storage_dir: &str, file_name: &str) -> Result<String, AppError> {
+    let safe_file_name = sanitize_filename::sanitize(file_name);
+    let mut final_filename = safe_file_name.clone();
+    let mut counter = 1;
+    let mut final_storage_path = format!("{storage_dir}{final_filename}");
+
+    while StdPath::new(&final_storage_path).exists() {
+        let path = StdPath::new(&safe_file_name);
+        let stem = path
+            .file_stem()
+            .and_then(|s| s.to_str())
+            .unwrap_or(&safe_file_name);
+        let extension = path.extension().and_then(|s| s.to_str()).unwrap_or("");
+
+        final_filename = if extension.is_empty() {
+            format!("{}({})", stem, counter)
+        } else {
+            format!("{}({}).{}", stem, counter, extension)
+        };
+        final_storage_path = format!("{storage_dir}{final_filename}");
+        counter += 1;
+
+        if counter > 1000 {
+            return Err(AppError::internal("Too many files with the same name"));
+        }
+    }
+
+    Ok(final_storage_path)
+}
+
+async fn create_content_record(
+    repository: &RoomContentRepository,
+    room_id: i64,
+    file: &UploadFileDescriptor,
+    final_storage_path: &str,
+) -> Result<RoomContent, AppError> {
+    repository
+        .create(&build_room_content(room_id, file, final_storage_path))
+        .await
+        .map_err(|e| AppError::internal(format!("创建内容记录失败：{}", e)))
+}
+
+fn build_room_content(
+    room_id: i64,
+    file: &UploadFileDescriptor,
+    final_storage_path: &str,
+) -> RoomContent {
+    let now = chrono::Utc::now().naive_utc();
+    RoomContent {
+        id: None,
+        room_id,
+        content_type: ContentType::File,
+        text: None,
+        url: Some(file.name.clone()),
+        path: Some(final_storage_path.to_string()),
+        file_name: Some(file.name.clone()),
+        size: Some(file.size),
+        mime_type: Some(
+            file.mime
+                .clone()
+                .unwrap_or_else(|| "application/octet-stream".to_string()),
+        ),
+        sequence_number: 0,
+        created_at: now,
+        updated_at: now,
+    }
+}
+
+async fn update_room_size(
+    repository: &RoomRepository,
+    room: &Room,
+    reserved_size: i64,
+    file_size: i64,
+) -> Result<(), AppError> {
+    let mut updated_room = room.clone();
+    updated_room.current_size = (updated_room.current_size - reserved_size + file_size).max(0);
+    repository
+        .update(&updated_room)
+        .await
+        .map_err(|e| AppError::internal(format!("更新房间大小失败：{}", e)))?;
+    Ok(())
+}
+
+async fn cleanup_temp_dir(temp_dir: &str) {
+    if let Err(e) = fs::remove_dir_all(temp_dir).await {
+        logrs::error!("清理临时文件失败：{}", e);
     }
 }
