+++ crates/board/src/tests/mod.rs
@@ -1,2 +1,3 @@
+mod cfg_service;
 mod room_gc_service;
 mod rooms_issue_token;
