+++ crates/board/tests/api_integration/mod.rs
@@ -1,0 +1,4 @@
+mod basic;
+mod content_flow;
+mod permissions;
+mod tokens;
