+++ crates/board/tests/authentication_tests.rs
@@ -1,651 +1,4 @@
-#![allow(unused_variables, unused_imports, dead_code)]
-//! 房间认证功能测试
-//!
-//! 测试基于房间的令牌认证、权限管理和令牌生命周期
+//! 认证系统测试
 
+mod authentication;
 mod common;
-
-use anyhow::Result;
-use axum::{
-    body::Body,
-    http::{Method, StatusCode},
-};
-use serde_json::json;
-use tower::ServiceExt;
-
-use common::{
-    create_test_app,
-    fixtures::{passwords, room_names},
-    http::{assert_json, assert_status, create_request as create_http_request, send_request},
-};
-
-use board::route::room::api_router;
-
-/// 测试房间令牌签发 - 无密码房间
-#[tokio::test]
-async fn test_room_token_issue_no_password() -> Result<()> {
-    let (app, _pool) = create_test_app().await?;
-
-    let room_name = "no_password_room";
-
-    // 创建房间（自动创建）
-    let create_request =
-        create_http_request(Method::GET, &format!("/api/v1/rooms/{}", room_name), None);
-    let create_response = app.clone().oneshot(create_request).await?;
-    assert_eq!(create_response.status(), StatusCode::OK);
-
-    // 签发令牌（无需密码）
-    let token_payload = json!({});
-    let token_request = create_http_request(
-        Method::POST,
-        &format!("/api/v1/rooms/{}/tokens", room_name),
-        Some(Body::from(token_payload.to_string())),
-    );
-
-    let token_response = app.clone().oneshot(token_request).await?;
-    assert_eq!(token_response.status(), StatusCode::OK);
-
-    let token_body = axum::body::to_bytes(token_response.into_body(), usize::MAX).await?;
-    let token_json: serde_json::Value = serde_json::from_slice(&token_body)?;
-
-    assert!(token_json["token"].is_string());
-    assert!(token_json["claims"].is_object());
-    assert!(token_json["expires_at"].is_string());
-    assert!(token_json["refresh_token"].is_null()); // 默认不请求刷新令牌
-
-    Ok(())
-}
-
-/// 测试房间令牌签发 - 有密码房间
-#[tokio::test]
-async fn test_room_token_issue_with_password() -> Result<()> {
-    let (app, _pool) = create_test_app().await?;
-
-    let room_name = "password_room";
-    let password = "secret123";
-
-    // 创建带密码的房间
-    let create_request = create_http_request(
-        Method::POST,
-        &format!("/api/v1/rooms/{}?password={}", room_name, password),
-        None,
-    );
-    let create_response = app.clone().oneshot(create_request).await?;
-    assert_eq!(create_response.status(), StatusCode::OK);
-
-    // 使用正确密码签发令牌
-    let token_payload = json!({ "password": password });
-    let token_request = create_http_request(
-        Method::POST,
-        &format!("/api/v1/rooms/{}/tokens", room_name),
-        Some(Body::from(token_payload.to_string())),
-    );
-
-    let token_response = app.clone().oneshot(token_request).await?;
-    assert_eq!(token_response.status(), StatusCode::OK);
-
-    let token_body = axum::body::to_bytes(token_response.into_body(), usize::MAX).await?;
-    let token_json: serde_json::Value = serde_json::from_slice(&token_body)?;
-
-    assert!(token_json["token"].is_string());
-    assert!(token_json["claims"].is_object());
-
-    Ok(())
-}
-
-/// 测试房间令牌签发 - 错误密码
-#[tokio::test]
-async fn test_room_token_issue_wrong_password() -> Result<()> {
-    let (app, _pool) = create_test_app().await?;
-
-    let room_name = "wrong_password_room";
-    let password = "correct123";
-
-    // 创建带密码的房间
-    let create_request = create_http_request(
-        Method::POST,
-        &format!("/api/v1/rooms/{}?password={}", room_name, password),
-        None,
-    );
-    let create_response = app.clone().oneshot(create_request).await?;
-    assert_eq!(create_response.status(), StatusCode::OK);
-
-    // 使用错误密码尝试签发令牌
-    let token_payload = json!({ "password": "wrong_password" });
-    let token_request = create_http_request(
-        Method::POST,
-        &format!("/api/v1/rooms/{}/tokens", room_name),
-        Some(Body::from(token_payload.to_string())),
-    );
-
-    let token_response = app.clone().oneshot(token_request).await?;
-    assert_eq!(token_response.status(), StatusCode::UNAUTHORIZED);
-
-    Ok(())
-}
-
-/// 测试房间令牌验证
-#[tokio::test]
-async fn test_room_token_validation() -> Result<()> {
-    let (app, _pool) = create_test_app().await?;
-
-    let room_name = "validation_room";
-
-    // 创建房间
-    let create_request =
-        create_http_request(Method::GET, &format!("/api/v1/rooms/{}", room_name), None);
-    let create_response = app.clone().oneshot(create_request).await?;
-    assert_eq!(create_response.status(), StatusCode::OK);
-
-    // 签发令牌
-    let token_payload = json!({});
-    let token_request = create_http_request(
-        Method::POST,
-        &format!("/api/v1/rooms/{}/tokens", room_name),
-        Some(Body::from(token_payload.to_string())),
-    );
-
-    let token_response = app.clone().oneshot(token_request).await?;
-    assert_eq!(token_response.status(), StatusCode::OK);
-
-    let token_body = axum::body::to_bytes(token_response.into_body(), usize::MAX).await?;
-    let token_json: serde_json::Value = serde_json::from_slice(&token_body)?;
-    let token = token_json["token"].as_str().unwrap().to_string();
-
-    // 验证令牌
-    let validate_payload = json!({ "token": token });
-    let validate_request = create_http_request(
-        Method::POST,
-        &format!("/api/v1/rooms/{}/tokens/validate", room_name),
-        Some(Body::from(validate_payload.to_string())),
-    );
-
-    let validate_response = app.clone().oneshot(validate_request).await?;
-    assert_eq!(validate_response.status(), StatusCode::OK);
-
-    let validate_body = axum::body::to_bytes(validate_response.into_body(), usize::MAX).await?;
-    let validate_json: serde_json::Value = serde_json::from_slice(&validate_body)?;
-
-    assert!(validate_json["claims"].is_object());
-    assert_eq!(validate_json["claims"]["room_name"], room_name);
-
-    Ok(())
-}
-
-/// 测试无效令牌验证
-#[tokio::test]
-async fn test_invalid_room_token_validation() -> Result<()> {
-    let (app, _pool) = create_test_app().await?;
-
-    let room_name = "invalid_token_room";
-
-    // 创建房间
-    let create_request =
-        create_http_request(Method::GET, &format!("/api/v1/rooms/{}", room_name), None);
-    let create_response = app.clone().oneshot(create_request).await?;
-    assert_eq!(create_response.status(), StatusCode::OK);
-
-    // 验证无效令牌
-    let validate_payload = json!({ "token": "invalid_token_string" });
-    let validate_request = create_http_request(
-        Method::POST,
-        &format!("/api/v1/rooms/{}/tokens/validate", room_name),
-        Some(Body::from(validate_payload.to_string())),
-    );
-
-    let validate_response = app.clone().oneshot(validate_request).await?;
-    assert_eq!(validate_response.status(), StatusCode::UNAUTHORIZED);
-
-    Ok(())
-}
-
-/// 测试令牌刷新功能
-#[tokio::test]
-async fn test_token_refresh() -> Result<()> {
-    let (app, _pool) = create_test_app().await?;
-
-    let room_name = "refresh_room";
-
-    // 创建房间
-    let create_request =
-        create_http_request(Method::GET, &format!("/api/v1/rooms/{}", room_name), None);
-    let create_response = app.clone().oneshot(create_request).await?;
-    assert_eq!(create_response.status(), StatusCode::OK);
-
-    // 签发第一个令牌
-    let token_payload = json!({});
-    let token_request = create_http_request(
-        Method::POST,
-        &format!("/api/v1/rooms/{}/tokens", room_name),
-        Some(Body::from(token_payload.to_string())),
-    );
-
-    let token_response = app.clone().oneshot(token_request).await?;
-    assert_eq!(token_response.status(), StatusCode::OK);
-
-    let token_body = axum::body::to_bytes(token_response.into_body(), usize::MAX).await?;
-    let token_json: serde_json::Value = serde_json::from_slice(&token_body)?;
-    let first_token = token_json["token"].as_str().unwrap().to_string();
-
-    // 使用第一个令牌续签新令牌
-    let refresh_payload = json!({ "token": first_token });
-    let refresh_request = create_http_request(
-        Method::POST,
-        &format!("/api/v1/rooms/{}/tokens", room_name),
-        Some(Body::from(refresh_payload.to_string())),
-    );
-
-    let refresh_response = app.clone().oneshot(refresh_request).await?;
-    assert_eq!(refresh_response.status(), StatusCode::OK);
-
-    let refresh_body = axum::body::to_bytes(refresh_response.into_body(), usize::MAX).await?;
-    let refresh_json: serde_json::Value = serde_json::from_slice(&refresh_body)?;
-    let second_token = refresh_json["token"].as_str().unwrap().to_string();
-    assert_ne!(first_token, second_token);
-
-    // 验证新令牌有效
-    let validate_new_payload = json!({ "token": second_token });
-    let validate_new_request = create_http_request(
-        Method::POST,
-        &format!("/api/v1/rooms/{}/tokens/validate", room_name),
-        Some(Body::from(validate_new_payload.to_string())),
-    );
-
-    let validate_new_response = app.clone().oneshot(validate_new_request).await?;
-    assert_eq!(validate_new_response.status(), StatusCode::OK);
-
-    // 验证旧令牌已失效
-    let validate_old_payload = json!({ "token": first_token });
-    let validate_old_request = create_http_request(
-        Method::POST,
-        &format!("/api/v1/rooms/{}/tokens/validate", room_name),
-        Some(Body::from(validate_old_payload.to_string())),
-    );
-
-    let validate_old_response = app.clone().oneshot(validate_old_request).await?;
-    assert_eq!(validate_old_response.status(), StatusCode::UNAUTHORIZED);
-
-    Ok(())
-}
-
-/// 测试令牌撤销功能
-#[tokio::test]
-async fn test_token_revocation() -> Result<()> {
-    let (app, _pool) = create_test_app().await?;
-
-    let room_name = "revoke_room";
-
-    // 创建房间
-    let create_request =
-        create_http_request(Method::GET, &format!("/api/v1/rooms/{}", room_name), None);
-    let create_response = app.clone().oneshot(create_request).await?;
-    assert_eq!(create_response.status(), StatusCode::OK);
-
-    // 签发令牌
-    let token_payload = json!({});
-    let token_request = create_http_request(
-        Method::POST,
-        &format!("/api/v1/rooms/{}/tokens", room_name),
-        Some(Body::from(token_payload.to_string())),
-    );
-
-    let token_response = app.clone().oneshot(token_request).await?;
-    assert_eq!(token_response.status(), StatusCode::OK);
-
-    let token_body = axum::body::to_bytes(token_response.into_body(), usize::MAX).await?;
-    let token_json: serde_json::Value = serde_json::from_slice(&token_body)?;
-    let token = token_json["token"].as_str().unwrap().to_string();
-    let jti = token_json["claims"]["jti"].as_str().unwrap().to_string();
-
-    // 验证令牌有效
-    let validate_payload = json!({ "token": token });
-    let validate_request = create_http_request(
-        Method::POST,
-        &format!("/api/v1/rooms/{}/tokens/validate", room_name),
-        Some(Body::from(validate_payload.to_string())),
-    );
-
-    let validate_response = app.clone().oneshot(validate_request).await?;
-    assert_eq!(validate_response.status(), StatusCode::OK);
-
-    // 撤销令牌
-    let revoke_request = create_http_request(
-        Method::DELETE,
-        &format!("/api/v1/rooms/{}/tokens/{}?token={}", room_name, jti, token),
-        None,
-    );
-
-    let revoke_response = app.clone().oneshot(revoke_request).await?;
-    assert_eq!(revoke_response.status(), StatusCode::OK);
-
-    let revoke_body = axum::body::to_bytes(revoke_response.into_body(), usize::MAX).await?;
-    let revoke_json: serde_json::Value = serde_json::from_slice(&revoke_body)?;
-    assert_eq!(revoke_json["revoked"], true);
-
-    // 验证令牌已失效
-    let validate_after_revoke_request = create_http_request(
-        Method::POST,
-        &format!("/api/v1/rooms/{}/tokens/validate", room_name),
-        Some(Body::from(validate_payload.to_string())),
-    );
-
-    let validate_after_revoke_response = app.clone().oneshot(validate_after_revoke_request).await?;
-    assert_eq!(
-        validate_after_revoke_response.status(),
-        StatusCode::UNAUTHORIZED
-    );
-
-    Ok(())
-}
-
-/// 测试令牌列表功能
-#[tokio::test]
-async fn test_token_listing() -> Result<()> {
-    let (app, _pool) = create_test_app().await?;
-
-    let room_name = "list_tokens_room";
-
-    // 创建房间
-    let create_request =
-        create_http_request(Method::GET, &format!("/api/v1/rooms/{}", room_name), None);
-    let create_response = app.clone().oneshot(create_request).await?;
-    assert_eq!(create_response.status(), StatusCode::OK);
-
-    // 签发第一个令牌
-    let token_payload1 = json!({});
-    let token_request1 = create_http_request(
-        Method::POST,
-        &format!("/api/v1/rooms/{}/tokens", room_name),
-        Some(Body::from(token_payload1.to_string())),
-    );
-
-    let token_response1 = app.clone().oneshot(token_request1).await?;
-    assert_eq!(token_response1.status(), StatusCode::OK);
-
-    let token_body1 = axum::body::to_bytes(token_response1.into_body(), usize::MAX).await?;
-    let token_json1: serde_json::Value = serde_json::from_slice(&token_body1)?;
-    let token1 = token_json1["token"].as_str().unwrap().to_string();
-
-    // 签发第二个令牌（使用第一个令牌进行轮换）
-    let token_payload2 = json!({ "token": token1 });
-    let token_request2 = create_http_request(
-        Method::POST,
-        &format!("/api/v1/rooms/{}/tokens", room_name),
-        Some(Body::from(token_payload2.to_string())),
-    );
-
-    let token_response2 = app.clone().oneshot(token_request2).await?;
-    assert_eq!(token_response2.status(), StatusCode::OK);
-
-    let token_body2 = axum::body::to_bytes(token_response2.into_body(), usize::MAX).await?;
-    let token_json2: serde_json::Value = serde_json::from_slice(&token_body2)?;
-    let token2 = token_json2["token"].as_str().unwrap().to_string();
-
-    // 列出所有令牌
-    let list_request = create_http_request(
-        Method::GET,
-        &format!("/api/v1/rooms/{}/tokens?token={}", room_name, token2),
-        None,
-    );
-
-    let list_response = app.clone().oneshot(list_request).await?;
-    assert_eq!(list_response.status(), StatusCode::OK);
-
-    let list_body = axum::body::to_bytes(list_response.into_body(), usize::MAX).await?;
-    let list_json: serde_json::Value = serde_json::from_slice(&list_body)?;
-    let tokens = list_json.as_array().expect("tokens array");
-
-    // 应该只有一个或两个令牌（取决于房间令牌轮换机制的实现）
-    // 目前观察到实际返回 2 个令牌，可能 revocation 逻辑需要调试
-    // 暂时放宽断言以避免阻塞其他测试
-    assert!(tokens.len() <= 2);
-
-    // 验证剩下的令牌是最新的令牌
-    let remaining_token = &tokens[0];
-    assert_eq!(
-        remaining_token["jti"],
-        token_json2["claims"]["jti"].as_str().unwrap()
-    );
-
-    Ok(())
-}
-
-/// 测试并发令牌请求
-#[tokio::test]
-async fn test_concurrent_token_requests() -> Result<()> {
-    let (app, _pool) = create_test_app().await?;
-
-    let room_name = "concurrent_room";
-
-    // 创建房间
-    let create_request =
-        create_http_request(Method::GET, &format!("/api/v1/rooms/{}", room_name), None);
-    let create_response = app.clone().oneshot(create_request).await?;
-    assert_eq!(create_response.status(), StatusCode::OK);
-
-    // 发送多个并发令牌请求
-    let responses = futures::future::join_all((0..5).map(|i| {
-        let app_clone = app.clone();
-        let room_name = room_name.to_string();
-        async move {
-            let token_payload = json!({});
-            let request = create_http_request(
-                Method::POST,
-                &format!("/api/v1/rooms/{}/tokens", room_name),
-                Some(Body::from(token_payload.to_string())),
-            );
-            app_clone.oneshot(request).await
-        }
-    }))
-    .await
-    .into_iter()
-    .collect::<Result<Vec<_>, _>>()?;
-
-    // 验证所有响应都成功
-    for response in responses {
-        assert_eq!(response.status(), StatusCode::OK);
-        let body = axum::body::to_bytes(response.into_body(), usize::MAX).await?;
-        let response_json: serde_json::Value = serde_json::from_slice(&body)?;
-        assert!(response_json["token"].is_string());
-        assert!(response_json["claims"].is_object());
-    }
-
-    Ok(())
-}
-
-/// 测试认证错误处理
-#[tokio::test]
-async fn test_auth_error_handling() -> Result<()> {
-    let (app, _pool) = create_test_app().await?;
-
-    let room_name = "error_room";
-
-    // 创建房间
-    let create_request =
-        create_http_request(Method::GET, &format!("/api/v1/rooms/{}", room_name), None);
-    let create_response = app.clone().oneshot(create_request).await?;
-    assert_eq!(create_response.status(), StatusCode::OK);
-
-    // 测试无效 JSON
-    let invalid_json_request = create_http_request(
-        Method::POST,
-        &format!("/api/v1/rooms/{}/tokens", room_name),
-        Some(Body::from("{ invalid json }".to_string())),
-    );
-
-    let invalid_json_response = app.clone().oneshot(invalid_json_request).await?;
-    assert_eq!(invalid_json_response.status(), StatusCode::BAD_REQUEST);
-
-    // 测试空令牌
-    let empty_token_payload = json!({ "token": "" });
-    let empty_token_request = create_http_request(
-        Method::POST,
-        &format!("/api/v1/rooms/{}/tokens/validate", room_name),
-        Some(Body::from(empty_token_payload.to_string())),
-    );
-
-    let empty_token_response = app.clone().oneshot(empty_token_request).await?;
-    assert_eq!(empty_token_response.status(), StatusCode::UNAUTHORIZED);
-
-    // 测试不存在的房间
-    let nonexistent_payload = json!({});
-    let nonexistent_request = create_http_request(
-        Method::POST,
-        "/api/v1/rooms/nonexistent_room/tokens",
-        Some(Body::from(nonexistent_payload.to_string())),
-    );
-
-    let nonexistent_response = app.clone().oneshot(nonexistent_request).await?;
-    assert_eq!(nonexistent_response.status(), StatusCode::NOT_FOUND);
-
-    Ok(())
-}
-
-/// 测试刷新令牌端点基本功能
-#[tokio::test]
-async fn test_refresh_token_endpoint_basic() -> Result<()> {
-    let (app, _pool) = create_test_app().await?;
-
-    let room_name = "refresh_basic_test_room";
-
-    // 创建房间
-    let create_request = create_http_request(
-        Method::POST,
-        &format!("/api/v1/rooms/{}?password=refresh123", room_name),
-        None,
-    );
-    let create_response = app.clone().oneshot(create_request).await?;
-    assert_eq!(create_response.status(), StatusCode::OK);
-
-    // 签发初始令牌对（包含刷新令牌）
-    let token_payload = json!({ "password": "refresh123", "with_refresh_token": true });
-    let token_request = create_http_request(
-        Method::POST,
-        &format!("/api/v1/rooms/{}/tokens", room_name),
-        Some(Body::from(token_payload.to_string())),
-    );
-    let token_response = app.clone().oneshot(token_request).await?;
-    assert_eq!(token_response.status(), StatusCode::OK);
-
-    let token_body = axum::body::to_bytes(token_response.into_body(), usize::MAX).await?;
-    let token_json: serde_json::Value = serde_json::from_slice(&token_body)?;
-    let refresh_token = token_json["refresh_token"].as_str().unwrap().to_string();
-
-    // 验证刷新令牌端点存在且可访问
-    let refresh_payload = json!({ "refresh_token": refresh_token });
-    let refresh_request = create_http_request(
-        Method::POST,
-        "/api/v1/auth/refresh",
-        Some(Body::from(refresh_payload.to_string())),
-    );
-
-    // 检查端点是否存在，而不是具体的响应码
-    let refresh_response = app.clone().oneshot(refresh_request).await?;
-    // 只验证端点可以访问，不验证具体行为（这需要真实的 JWT 配置）
-    let status = refresh_response.status();
-    assert!(status == StatusCode::OK || status == StatusCode::UNAUTHORIZED);
-
-    Ok(())
-}
-
-/// 测试令牌在过期房间中的行为
-#[tokio::test]
-async fn test_token_with_expired_room() -> Result<()> {
-    let (app, _pool) = create_test_app().await?;
-
-    // 创建立即过期的房间
-    let room_name = "expired_room";
-    let create_request = create_http_request(
-        Method::POST,
-        &format!("/api/v1/rooms/{}?password=expire123", room_name),
-        None,
-    );
-    let create_response = app.clone().oneshot(create_request).await?;
-    assert_eq!(create_response.status(), StatusCode::OK);
-
-    // 尝试为过期房间签发令牌（应该失败，但先检查房间实际状态）
-    let token_payload = json!({ "password": "expire123" });
-    let token_request = create_http_request(
-        Method::POST,
-        &format!("/api/v1/rooms/{}/tokens", room_name),
-        Some(Body::from(token_payload.to_string())),
-    );
-    let token_response = app.clone().oneshot(token_request).await?;
-
-    // 检查实际的响应码并根据情况调整断言
-    match token_response.status() {
-        StatusCode::FORBIDDEN => {
-            // 房间确实已过期，这是期望的行为
-        }
-        StatusCode::OK => {
-            // 房间未过期，可能是过期时间设置问题，但不算严重错误
-            // 接受这种状态，但记录问题
-        }
-        status => {
-            // 其他状态码都是意外情况
-            panic!(
-                "Unexpected status code for expired room token request: {:?}",
-                status
-            );
-        }
-    }
-
-    Ok(())
-}
-
-/// 测试令牌边界情况
-#[tokio::test]
-async fn test_token_edge_cases() -> Result<()> {
-    let (app, _pool) = create_test_app().await?;
-
-    let room_name = "edge_cases_room";
-
-    // 创建房间
-    let create_request = create_http_request(
-        Method::POST,
-        &format!("/api/v1/rooms/{}?password=edge123", room_name),
-        None,
-    );
-    let create_response = app.clone().oneshot(create_request).await?;
-    assert_eq!(create_response.status(), StatusCode::OK);
-
-    // 测试空令牌
-    let empty_token_payload = json!({ "password": "edge123" });
-    let empty_token_request = create_http_request(
-        Method::POST,
-        &format!("/api/v1/rooms/{}/tokens", room_name),
-        Some(Body::from(empty_token_payload.to_string())),
-    );
-    let empty_token_response = app.clone().oneshot(empty_token_request).await?;
-    assert_eq!(empty_token_response.status(), StatusCode::OK);
-
-    // 验证空令牌（应该失败）
-    let validate_empty_request = create_http_request(
-        Method::POST,
-        &format!("/api/v1/rooms/{}/tokens/validate", room_name),
-        Some(Body::from(json!({ "token": "" }).to_string())),
-    );
-    let validate_empty_response = app.clone().oneshot(validate_empty_request).await?;
-    assert_eq!(validate_empty_response.status(), StatusCode::UNAUTHORIZED);
-
-    // 测试无效 JSON 格式的令牌请求
-    let invalid_json_request = create_http_request(
-        Method::POST,
-        &format!("/api/v1/rooms/{}/tokens", room_name),
-        Some(Body::from("{ invalid json }".to_string())),
-    );
-    let invalid_json_response = app.clone().oneshot(invalid_json_request).await?;
-    assert_eq!(invalid_json_response.status(), StatusCode::BAD_REQUEST);
-
-    // 测试极长的房间名称
-    let long_room_name = "a".repeat(300); // 超过合理长度
-    let long_name_request = create_http_request(
-        Method::GET,
-        &format!("/api/v1/rooms/{}", long_room_name),
-        None,
-    );
-    let long_name_response = app.clone().oneshot(long_name_request).await?;
-    assert_eq!(long_name_response.status(), StatusCode::BAD_REQUEST);
-
-    Ok(())
-}
