+++ crates/board/tests/storage_system_tests.rs
@@ -1,660 +1,4 @@
-#![allow(unused_variables, unused_imports, dead_code)]
-//! 存储系统功能测试
-//!
-//! 测试内容管理、上传预留、分块合并、存储路径等存储相关功能
+//! 存储系统测试
 
 mod common;
-
-use anyhow::Result;
-use axum::{
-    body::Body,
-    http::{Method, StatusCode},
-};
-use serde_json::json;
-use tower::ServiceExt;
-
-use common::{
-    create_test_app,
-    fixtures::{file_sizes, filenames, passwords, room_names},
-    http::{assert_json, assert_status, create_request as create_http_request, send_request},
-};
-
-/// 测试内容管理 - 文本内容
-#[tokio::test]
-async fn test_content_management_text() -> Result<()> {
-    let (app, _pool) = create_test_app().await?;
-
-    let room_name = "text_content_test_room";
-    let password = "text123";
-
-    // 创建带密码的房间
-    let create_request = create_http_request(
-        Method::POST,
-        &format!("/api/v1/rooms/{}?password={}", room_name, password),
-        None,
-    );
-    let create_response = app.clone().oneshot(create_request).await?;
-    assert_eq!(create_response.status(), StatusCode::OK);
-
-    // 签发令牌
-    let token_payload = json!({ "password": password });
-    let token_request = create_http_request(
-        Method::POST,
-        &format!("/api/v1/rooms/{}/tokens", room_name),
-        Some(Body::from(token_payload.to_string())),
-    );
-    let token_response = app.clone().oneshot(token_request).await?;
-    assert_eq!(token_response.status(), StatusCode::OK);
-
-    let token_body = axum::body::to_bytes(token_response.into_body(), usize::MAX).await?;
-    let token_json: serde_json::Value = serde_json::from_slice(&token_body)?;
-    let token = token_json["token"].as_str().unwrap().to_string();
-
-    // 更新房间内容 - 文本
-    let content_payload = json!({
-        "text": "Hello, this is test content!",
-        "urls": []
-    });
-    let content_request = create_http_request(
-        Method::PUT,
-        &format!("/api/v1/rooms/{}/content?token={}", room_name, token),
-        Some(Body::from(content_payload.to_string())),
-    );
-    let content_response = app.clone().oneshot(content_request).await?;
-
-    // 内容更新端点可能实现或未实现
-    let content_status = content_response.status();
-    println!("Content update status: {}", content_status);
-
-    if content_status == StatusCode::OK {
-        // 如果端点实现，继续测试内容获取
-    } else {
-        // 如果端点未实现，跳过验证
-        println!("Content update endpoint not implemented, skipping verification");
-        return Ok(());
-    }
-
-    // 获取房间内容
-    let get_request = create_http_request(
-        Method::GET,
-        &format!("/api/v1/rooms/{}?include_content=true", room_name),
-        None,
-    );
-    let get_response = app.clone().oneshot(get_request).await?;
-    assert_eq!(get_response.status(), StatusCode::OK);
-
-    let get_body = axum::body::to_bytes(get_response.into_body(), usize::MAX).await?;
-    let get_json: serde_json::Value = serde_json::from_slice(&get_body)?;
-
-    // 验证内容存在
-    assert!(get_json["content"].is_object());
-    assert_eq!(get_json["content"]["text"], "Hello, this is test content!");
-
-    Ok(())
-}
-
-/// 测试内容管理 - URL 内容
-#[tokio::test]
-async fn test_content_management_urls() -> Result<()> {
-    let (app, _pool) = create_test_app().await?;
-
-    let room_name = "url_content_test_room";
-
-    // 创建房间
-    let create_request =
-        create_http_request(Method::GET, &format!("/api/v1/rooms/{}", room_name), None);
-    let create_response = app.clone().oneshot(create_request).await?;
-    assert_eq!(create_response.status(), StatusCode::OK);
-
-    // 签发令牌
-    let token_payload = json!({});
-    let token_request = create_http_request(
-        Method::POST,
-        &format!("/api/v1/rooms/{}/tokens", room_name),
-        Some(Body::from(token_payload.to_string())),
-    );
-    let token_response = app.clone().oneshot(token_request).await?;
-    assert_eq!(token_response.status(), StatusCode::OK);
-
-    let token_body = axum::body::to_bytes(token_response.into_body(), usize::MAX).await?;
-    let token_json: serde_json::Value = serde_json::from_slice(&token_body)?;
-    let token = token_json["token"].as_str().unwrap().to_string();
-
-    // 更新房间内容 - URLs
-    let content_payload = json!({
-        "text": "",
-        "urls": [
-            "https://example.com/image1.jpg",
-            "https://example.com/image2.png"
-        ]
-    });
-    let content_request = create_http_request(
-        Method::PUT,
-        &format!("/api/v1/rooms/{}/content?token={}", room_name, token),
-        Some(Body::from(content_payload.to_string())),
-    );
-    let content_response = app.clone().oneshot(content_request).await?;
-
-    // 内容更新端点可能实现或未实现
-    let content_status = content_response.status();
-    println!("URL content update status: {}", content_status);
-
-    if content_status == StatusCode::OK {
-        // 如果端点实现，继续测试内容获取
-    } else {
-        println!("Content update endpoint not implemented, skipping verification");
-        return Ok(());
-    }
-
-    // 获取房间内容
-    let get_request = create_http_request(
-        Method::GET,
-        &format!("/api/v1/rooms/{}?include_content=true", room_name),
-        None,
-    );
-    let get_response = app.oneshot(get_request).await?;
-    assert_eq!(get_response.status(), StatusCode::OK);
-
-    let get_body = axum::body::to_bytes(get_response.into_body(), usize::MAX).await?;
-    let get_json: serde_json::Value = serde_json::from_slice(&get_body)?;
-
-    // 验证 URLs 存在
-    assert!(get_json["content"].is_object());
-    assert_eq!(get_json["content"]["text"], "");
-    let urls = get_json["content"]["urls"].as_array().unwrap();
-    assert_eq!(urls.len(), 2);
-    assert_eq!(urls[0], "https://example.com/image1.jpg");
-    assert_eq!(urls[1], "https://example.com/image2.png");
-
-    Ok(())
-}
-
-/// 测试内容管理 - 混合内容
-#[tokio::test]
-async fn test_content_management_mixed() -> Result<()> {
-    let (app, _pool) = create_test_app().await?;
-
-    let room_name = "mixed_content_test_room";
-
-    // 创建房间
-    let create_request =
-        create_http_request(Method::GET, &format!("/api/v1/rooms/{}", room_name), None);
-    let create_response = app.clone().oneshot(create_request).await?;
-    assert_eq!(create_response.status(), StatusCode::OK);
-
-    // 签发令牌
-    let token_payload = json!({});
-    let token_request = create_http_request(
-        Method::POST,
-        &format!("/api/v1/rooms/{}/tokens", room_name),
-        Some(Body::from(token_payload.to_string())),
-    );
-    let token_response = app.clone().oneshot(token_request).await?;
-    assert_eq!(token_response.status(), StatusCode::OK);
-
-    let token_body = axum::body::to_bytes(token_response.into_body(), usize::MAX).await?;
-    let token_json: serde_json::Value = serde_json::from_slice(&token_body)?;
-    let token = token_json["token"].as_str().unwrap().to_string();
-
-    // 更新房间内容 - 混合内容
-    let content_payload = json!({
-        "text": "This is mixed content with images:",
-        "urls": [
-            "https://example.com/image1.jpg",
-            "https://example.com/image2.png"
-        ]
-    });
-    let content_request = create_http_request(
-        Method::PUT,
-        &format!("/api/v1/rooms/{}/content?token={}", room_name, token),
-        Some(Body::from(content_payload.to_string())),
-    );
-    let content_response = app.clone().oneshot(content_request).await?;
-
-    // 内容更新端点可能实现或未实现
-    let content_status = content_response.status();
-    println!("Mixed content update status: {}", content_status);
-
-    if content_status == StatusCode::OK {
-        // 如果端点实现，继续测试内容获取
-    } else {
-        println!("Content update endpoint not implemented, skipping verification");
-        return Ok(());
-    }
-
-    // 验证内容存储大小和 UTF-8 字节计算
-    let get_request = create_http_request(
-        Method::GET,
-        &format!("/api/v1/rooms/{}?include_content=true", room_name),
-        None,
-    );
-    let get_response = app.oneshot(get_request).await?;
-    assert_eq!(get_response.status(), StatusCode::OK);
-
-    let get_body = axum::body::to_bytes(get_response.into_body(), usize::MAX).await?;
-    let get_json: serde_json::Value = serde_json::from_slice(&get_body)?;
-
-    // 验证混合内容存在
-    assert!(get_json["content"].is_object());
-    assert_eq!(
-        get_json["content"]["text"],
-        "This is mixed content with images:"
-    );
-    let urls = get_json["content"]["urls"].as_array().unwrap();
-    assert_eq!(urls.len(), 2);
-
-    Ok(())
-}
-
-/// 测试内容管理错误处理
-#[tokio::test]
-async fn test_content_management_errors() -> Result<()> {
-    let (app, _pool) = create_test_app().await?;
-
-    let room_name = "error_content_test_room";
-
-    // 创建房间
-    let create_request =
-        create_http_request(Method::GET, &format!("/api/v1/rooms/{}", room_name), None);
-    let create_response = app.clone().oneshot(create_request).await?;
-    assert_eq!(create_response.status(), StatusCode::OK);
-
-    // 测试无令牌更新内容
-    let content_payload = json!({
-        "text": "This should fail",
-        "urls": []
-    });
-    let content_request = create_http_request(
-        Method::PUT,
-        &format!("/api/v1/rooms/{}/content", room_name),
-        Some(Body::from(content_payload.to_string())),
-    );
-    let content_response = app.clone().oneshot(content_request).await?;
-
-    // 内容更新端点可能未实现，所以可能是 404 或 401
-    let content_status = content_response.status();
-    assert!(content_status == StatusCode::UNAUTHORIZED || content_status == StatusCode::NOT_FOUND);
-
-    // 测试无效令牌更新内容
-    let invalid_content_request = create_http_request(
-        Method::PUT,
-        &format!("/api/v1/rooms/{}/content?token=invalid_token", room_name),
-        Some(Body::from(content_payload.to_string())),
-    );
-    let invalid_content_response = app.clone().oneshot(invalid_content_request).await?;
-
-    // 内容更新端点可能未实现，所以可能是 404 或 401
-    let invalid_status = invalid_content_response.status();
-    assert!(invalid_status == StatusCode::UNAUTHORIZED || invalid_status == StatusCode::NOT_FOUND);
-
-    // 测试不存在的房间
-    let nonexistent_room_request = create_http_request(
-        Method::PUT,
-        "/api/v1/rooms/nonexistent_room/content?token=dummy_token",
-        Some(Body::from(content_payload.to_string())),
-    );
-    let nonexistent_room_response = app.clone().oneshot(nonexistent_room_request).await?;
-
-    // 不存在房间的请求应该返回 404（无论端点是否实现）
-    assert_eq!(nonexistent_room_response.status(), StatusCode::NOT_FOUND);
-
-    Ok(())
-}
-
-/// 测试上传预留功能
-#[tokio::test]
-async fn test_upload_reservation() -> Result<()> {
-    let (app, _pool) = create_test_app().await?;
-
-    let room_name = "upload_reservation_test_room";
-
-    // 创建房间
-    let create_request =
-        create_http_request(Method::GET, &format!("/api/v1/rooms/{}", room_name), None);
-    let create_response = app.clone().oneshot(create_request).await?;
-    assert_eq!(create_response.status(), StatusCode::OK);
-
-    // 签发令牌
-    let token_payload = json!({});
-    let token_request = create_http_request(
-        Method::POST,
-        &format!("/api/v1/rooms/{}/tokens", room_name),
-        Some(Body::from(token_payload.to_string())),
-    );
-    let token_response = app.clone().oneshot(token_request).await?;
-    assert_eq!(token_response.status(), StatusCode::OK);
-
-    let token_body = axum::body::to_bytes(token_response.into_body(), usize::MAX).await?;
-    let token_json: serde_json::Value = serde_json::from_slice(&token_body)?;
-    let token = token_json["token"].as_str().unwrap().to_string();
-
-    // 预留上传
-    let reserve_payload = json!({
-        "files": [
-            {
-                "name": "test_file.txt",
-                "size": 1024,
-                "content_type": "text/plain"
-            },
-            {
-                "name": "test_image.jpg",
-                "size": 2048,
-                "content_type": "image/jpeg"
-            }
-        ]
-    });
-    let reserve_request = create_http_request(
-        Method::POST,
-        &format!(
-            "/api/v1/rooms/{}/uploads/reserve?token={}",
-            room_name, token
-        ),
-        Some(Body::from(reserve_payload.to_string())),
-    );
-    let reserve_response = app.clone().oneshot(reserve_request).await?;
-
-    // 上传预留可能成功或失败，取决于 API 实现
-    let status = reserve_response.status();
-    println!("Upload reservation status: {}", status);
-
-    // 验证端点可访问且返回合理的响应
-    assert!(
-        status == StatusCode::OK
-            || status == StatusCode::NOT_IMPLEMENTED
-            || status == StatusCode::BAD_REQUEST
-            || status == StatusCode::NOT_FOUND
-    );
-
-    Ok(())
-}
-
-/// 测试分块合并功能
-#[tokio::test]
-async fn test_chunk_merge() -> Result<()> {
-    let (app, _pool) = create_test_app().await?;
-
-    let room_name = "chunk_merge_test_room";
-
-    // 创建房间
-    let create_request =
-        create_http_request(Method::GET, &format!("/api/v1/rooms/{}", room_name), None);
-    let create_response = app.clone().oneshot(create_request).await?;
-    assert_eq!(create_response.status(), StatusCode::OK);
-
-    // 签发令牌
-    let token_payload = json!({});
-    let token_request = create_http_request(
-        Method::POST,
-        &format!("/api/v1/rooms/{}/tokens", room_name),
-        Some(Body::from(token_payload.to_string())),
-    );
-    let token_response = app.clone().oneshot(token_request).await?;
-    assert_eq!(token_response.status(), StatusCode::OK);
-
-    let token_body = axum::body::to_bytes(token_response.into_body(), usize::MAX).await?;
-    let token_json: serde_json::Value = serde_json::from_slice(&token_body)?;
-    let token = token_json["token"].as_str().unwrap().to_string();
-
-    // 测试分块合并（使用预留 ID 或上传令牌）
-    let merge_payload = json!({
-        "reservation_id": "test_reservation_id",
-        "final_hash": "dummy_hash_value"
-    });
-    let merge_request = create_http_request(
-        Method::POST,
-        &format!("/api/v1/rooms/{}/uploads/merge?token={}", room_name, token),
-        Some(Body::from(merge_payload.to_string())),
-    );
-    let merge_response = app.clone().oneshot(merge_request).await?;
-
-    // 分块合并可能成功或失败，取决于 API 实现
-    let status = merge_response.status();
-    println!("Chunk merge status: {}", status);
-
-    // 验证端点可访问且返回合理的响应
-    assert!(
-        status == StatusCode::OK
-            || status == StatusCode::NOT_FOUND
-            || status == StatusCode::BAD_REQUEST
-            || status == StatusCode::NOT_IMPLEMENTED
-    );
-
-    Ok(())
-}
-
-/// 测试存储路径管理
-#[tokio::test]
-async fn test_storage_paths() -> Result<()> {
-    let (app, _pool) = create_test_app().await?;
-
-    let room_name = "storage_path_test_room";
-
-    // 创建房间
-    let create_request =
-        create_http_request(Method::GET, &format!("/api/v1/rooms/{}", room_name), None);
-    let create_response = app.clone().oneshot(create_request).await?;
-    assert_eq!(create_response.status(), StatusCode::OK);
-
-    // 签发令牌
-    let token_payload = json!({});
-    let token_request = create_http_request(
-        Method::POST,
-        &format!("/api/v1/rooms/{}/tokens", room_name),
-        Some(Body::from(token_payload.to_string())),
-    );
-    let token_response = app.clone().oneshot(token_request).await?;
-    assert_eq!(token_response.status(), StatusCode::OK);
-
-    let token_body = axum::body::to_bytes(token_response.into_body(), usize::MAX).await?;
-    let token_json: serde_json::Value = serde_json::from_slice(&token_body)?;
-    let token = token_json["token"].as_str().unwrap().to_string();
-
-    // 测试获取存储路径信息
-    let paths_request = create_http_request(
-        Method::GET,
-        &format!("/api/v1/rooms/{}/storage/paths?token={}", room_name, token),
-        None,
-    );
-    let paths_response = app.clone().oneshot(paths_request).await?;
-
-    // 存储路径端点可能实现或未实现
-    let status = paths_response.status();
-    println!("Storage paths status: {}", status);
-
-    // 验证端点行为合理
-    assert!(
-        status == StatusCode::OK
-            || status == StatusCode::NOT_FOUND
-            || status == StatusCode::NOT_IMPLEMENTED
-    );
-
-    Ok(())
-}
-
-/// 测试文件元数据管理
-#[tokio::test]
-async fn test_file_metadata() -> Result<()> {
-    let (app, _pool) = create_test_app().await?;
-
-    let room_name = "file_metadata_test_room";
-
-    // 创建房间
-    let create_request =
-        create_http_request(Method::GET, &format!("/api/v1/rooms/{}", room_name), None);
-    let create_response = app.clone().oneshot(create_request).await?;
-    assert_eq!(create_response.status(), StatusCode::OK);
-
-    // 签发令牌
-    let token_payload = json!({});
-    let token_request = create_http_request(
-        Method::POST,
-        &format!("/api/v1/rooms/{}/tokens", room_name),
-        Some(Body::from(token_payload.to_string())),
-    );
-    let token_response = app.clone().oneshot(token_request).await?;
-    assert_eq!(token_response.status(), StatusCode::OK);
-
-    let token_body = axum::body::to_bytes(token_response.into_body(), usize::MAX).await?;
-    let token_json: serde_json::Value = serde_json::from_slice(&token_body)?;
-    let token = token_json["token"].as_str().unwrap().to_string();
-
-    // 测试文件元数据端点
-    let metadata_request = create_http_request(
-        Method::GET,
-        &format!("/api/v1/rooms/{}/storage/files?token={}", room_name, token),
-        None,
-    );
-    let metadata_response = app.clone().oneshot(metadata_request).await?;
-
-    // 文件元数据端点可能实现或未实现
-    let status = metadata_response.status();
-    println!("File metadata status: {}", status);
-
-    // 验证端点行为合理
-    assert!(
-        status == StatusCode::OK
-            || status == StatusCode::NOT_FOUND
-            || status == StatusCode::NOT_IMPLEMENTED
-    );
-
-    Ok(())
-}
-
-/// 测试存储清理功能
-#[tokio::test]
-async fn test_storage_cleanup() -> Result<()> {
-    let (app, _pool) = create_test_app().await?;
-
-    let room_name = "storage_cleanup_test_room";
-
-    // 创建房间
-    let create_request =
-        create_http_request(Method::GET, &format!("/api/v1/rooms/{}", room_name), None);
-    let create_response = app.clone().oneshot(create_request).await?;
-    assert_eq!(create_response.status(), StatusCode::OK);
-
-    // 签发令牌
-    let token_payload = json!({});
-    let token_request = create_http_request(
-        Method::POST,
-        &format!("/api/v1/rooms/{}/tokens", room_name),
-        Some(Body::from(token_payload.to_string())),
-    );
-    let token_response = app.clone().oneshot(token_request).await?;
-    assert_eq!(token_response.status(), StatusCode::OK);
-
-    let token_body = axum::body::to_bytes(token_response.into_body(), usize::MAX).await?;
-    let token_json: serde_json::Value = serde_json::from_slice(&token_body)?;
-    let token = token_json["token"].as_str().unwrap().to_string();
-
-    // 测试存储清理端点
-    let cleanup_request = create_http_request(
-        Method::DELETE,
-        &format!(
-            "/api/v1/rooms/{}/storage/cleanup?token={}",
-            room_name, token
-        ),
-        None,
-    );
-    let cleanup_response = app.clone().oneshot(cleanup_request).await?;
-
-    // 存储清理端点可能实现或未实现
-    let status = cleanup_response.status();
-    println!("Storage cleanup status: {}", status);
-
-    // 验证端点行为合理
-    assert!(
-        status == StatusCode::OK
-            || status == StatusCode::NOT_FOUND
-            || status == StatusCode::NOT_IMPLEMENTED
-            || status == StatusCode::NO_CONTENT
-    );
-
-    Ok(())
-}
-
-/// 测试存储权限验证
-#[tokio::test]
-async fn test_storage_permissions() -> Result<()> {
-    let (app, _pool) = create_test_app().await?;
-
-    let room_name = "storage_perms_test_room";
-    let password = "storage123";
-
-    // 创建带密码的房间
-    let create_request = create_http_request(
-        Method::POST,
-        &format!("/api/v1/rooms/{}?password={}", room_name, password),
-        None,
-    );
-    let create_response = app.clone().oneshot(create_request).await?;
-    assert_eq!(create_response.status(), StatusCode::OK);
-
-    // 测试无令牌访问存储功能
-    let content_request = create_http_request(
-        Method::PUT,
-        &format!("/api/v1/rooms/{}/content", room_name),
-        Some(Body::from(
-            json!({
-                "text": "test content",
-                "urls": []
-            })
-            .to_string(),
-        )),
-    );
-    let content_response = app.clone().oneshot(content_request).await?;
-
-    // 内容更新端点可能未实现，所以可能是 404 或 401
-    let content_status = content_response.status();
-    assert!(content_status == StatusCode::UNAUTHORIZED || content_status == StatusCode::NOT_FOUND);
-
-    // 测试错误密码令牌
-    let wrong_token_payload = json!({ "password": "wrong_password" });
-    let wrong_token_request = create_http_request(
-        Method::POST,
-        &format!("/api/v1/rooms/{}/tokens", room_name),
-        Some(Body::from(wrong_token_payload.to_string())),
-    );
-    let wrong_token_response = app.clone().oneshot(wrong_token_request).await?;
-    assert_eq!(wrong_token_response.status(), StatusCode::UNAUTHORIZED);
-
-    // 测试正确密码令牌
-    let correct_token_payload = json!({ "password": password });
-    let correct_token_request = create_http_request(
-        Method::POST,
-        &format!("/api/v1/rooms/{}/tokens", room_name),
-        Some(Body::from(correct_token_payload.to_string())),
-    );
-    let correct_token_response = app.clone().oneshot(correct_token_request).await?;
-    assert_eq!(correct_token_response.status(), StatusCode::OK);
-
-    let token_body = axum::body::to_bytes(correct_token_response.into_body(), usize::MAX).await?;
-    let token_json: serde_json::Value = serde_json::from_slice(&token_body)?;
-    let token = token_json["token"].as_str().unwrap().to_string();
-
-    // 使用正确令牌测试内容更新
-    let valid_content_request = create_http_request(
-        Method::PUT,
-        &format!("/api/v1/rooms/{}/content?token={}", room_name, token),
-        Some(Body::from(
-            json!({
-                "text": "valid content",
-                "urls": []
-            })
-            .to_string(),
-        )),
-    );
-    let valid_content_response = app.oneshot(valid_content_request).await?;
-
-    // 内容更新端点可能未实现，所以可能是 404 或 200
-    let valid_status = valid_content_response.status();
-    assert!(valid_status == StatusCode::OK || valid_status == StatusCode::NOT_FOUND);
-
-    if valid_status == StatusCode::OK {
-        println!("Content update with valid token succeeded");
-    } else {
-        println!("Content update endpoint not implemented");
-    }
-
-    Ok(())
-}
+mod storage_system;
